feat: adds a method to complete the task
- Adds a method to complete the task instead of removing it.

BREAKING CHANGE: On clicking on the checkbox, the task status is updated to '1', i.e., marked as completed instead of deleting from database.
Swiping from either side will continue to delete the task.

This change introduced a bug in the 'Search' screen as the 'Search' screen continue to show all the tasks (open & completed). So when a completed task shown in 'Search Results', unchecking it, may not mark it as 'Not completed'. This is still being worked on.