feat: adds a screen to see the completed tasks
- Adds a new activity to show the completed tasks
- Adds new option to allow opening the completed tasks view
- Adds string resources for completed tasks labels & related translations
- Fixes the warning in android manifest & adds the new activity definition
- Fixes the warning in menu, i.e., uses 'IfRoom' instead of 'always'

BREAKING CHANGE: This adds a new activity screen that shows the completed tasks that's launched by clicking on a button in the app bar. It only shows the completed tasks.
These tasks can be moved to 'Open/Not completed' stage on clicking the checkbox. It will show up in the main tasks screen as not completed.