feat: adds swipe feature to completed tasks screen
- Adds swipe feature to completed tasks screen
- Adds the snackbar with 'undo' actions
- Removes unnecessary toast

BREAKING CHANGE: The swipe feature is added to completed tasks screen to help the user clean up the tasks. The swipe feauture is in consistent with the main tasks screen with one exception. Swiping from left to right moves the task from 'Completed' to 'Open' status. Swiping from 'Right' to 'Left' deletes the task as similar to Main Tasks Screen.